Split out relation-related types.
